rename module gpu::vulkan to gpu::vk
